# api/src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Memes API
  version: "1.0.0"
  description: |
    REST API for browsing and searching scraped Reddit memes.
    Push changes to memes/ to trigger automatic re-indexing via GitHub Actions.

servers:
  - url: /

tags:
  - name: memes
    description: Read-only browsing and search
  - name: admin
    description: Owner-only operations (requires X-Api-Key header)

paths:

  /:
    get:
      tags: [memes]
      operationId: getStats
      summary: Global collection statistics
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /categories:
    get:
      tags: [memes]
      operationId: listCategories
      summary: All categories with meme counts and top score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

  /memes:
    get:
      tags: [memes]
      operationId: listMemes
      summary: Paginated list of memes with optional filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
        - name: subreddit
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, created_at, title]
            default: score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemePage"

  /memes/{category}:
    get:
      tags: [memes]
      operationId: listMemesByCategory
      summary: All memes in a specific category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, created_at, title]
            default: score
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemePage"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /memes/{category}/{slug}:
    get:
      tags: [memes]
      operationId: getMeme
      summary: Single meme by category and slug
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meme"
        "404":
          description: Meme not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search:
    get:
      tags: [memes]
      operationId: searchMemes
      summary: Full-text search across title and description (FTS5)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemePage"
        "400":
          description: Missing or empty q parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/reindex:
    post:
      tags: [admin]
      operationId: reindex
      summary: Re-scan memes/ directory and upsert all memes into SQLite
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Reindex completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReindexResult"
        "401":
          description: Missing or invalid X-Api-Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error during indexing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

  schemas:
    Stats:
      type: object
      properties:
        total_memes:
          type: integer
        total_categories:
          type: integer
        total_subreddits:
          type: integer
        top_category:
          type: string
        indexed_at:
          type: string
          format: date-time

    CategorySummary:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        top_score:
          type: integer

    Meme:
      type: object
      required: [slug, category, title]
      properties:
        slug:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        author:
          type: string
        subreddit:
          type: string
        score:
          type: integer
        created_at:
          type: string
          format: date-time
        source_url:
          type: string
        post_url:
          type: string
        image_path:
          type: string
        tags:
          type: array
          items:
            type: string

    MemePage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Meme"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    ReindexResult:
      type: object
      properties:
        indexed:
          type: integer
        duration_ms:
          type: integer
          format: int64
        errors:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string